+++ src/main/java/com/nubble/backend/post/service/PostInfoMapper.java
@@ -1,0 +1,19 @@
+package com.nubble.backend.post.service;
+
+import com.nubble.backend.post.domain.Post;
+import org.mapstruct.InjectionStrategy;
+import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
+import org.mapstruct.ReportingPolicy;
+
+@Mapper(
+        componentModel = "spring",
+        injectionStrategy = InjectionStrategy.CONSTRUCTOR,
+        unmappedTargetPolicy = ReportingPolicy.ERROR
+)
+public interface PostInfoMapper {
+
+    @Mapping(target = "userId", source = "user.id")
+    @Mapping(target = "boardId", source = "board.id")
+    PostInfo.PostDto toPostDto(Post post);
+}
