+++ src/main/java/com/nubble/backend/post/service/PostRepository.java
@@ -1,8 +1,10 @@
 package com.nubble.backend.post.service;
 
 import com.nubble.backend.post.domain.Post;
+import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
 
+    List<Post> findAllByBoardId(long boardId);
 }
